ai-service/requirements.txt
# AI & RAG Engine Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Gemini AI
google-generativeai==0.3.2

# Database
psycopg2-binary==2.9.9

# Utilities
numpy==1.26.3
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
